refactor: generatePrefixを別ファイルに切り出し
